+++ src/main/java/com/bookripple/api/infrastructure/email/AwsSesEmailSender.java
@@ -1,13 +1,11 @@
 package com.bookripple.api.infrastructure.email;
 
+import com.bookripple.api.global.auth.util.EmailSender;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
-
-import com.bookripple.api.domain.auth.util.EmailSender;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.services.ses.SesClient;
 import software.amazon.awssdk.services.ses.model.Body;
 import software.amazon.awssdk.services.ses.model.Content;
